refactor: move ClientPrincipal type to /types
